requirements.txt
fastapi>=0.128.0
uvicorn>=0.40.0
requests>=2.32.5
pandas>=2.0.0,<3.0.0
duckdb>=1.4.4
pydantic>=2.12.5          # Updated to latest stable in 2.x series
python-dotenv>=1.2.1     # Updated to latest stable
google-generativeai>=0.3.0